requirements.txt
pyaudio
wave
python-dotenv
fastapi
uvicorn
faster-whisper

langchain
langchain-groq


pymongo

